Refactor logging implementations and add optional functionality
The logging configurations for RestTemplate and OkHttp have been updated to allow optional functionality with the creation of loggingDisabled and contextHeadersDisabled properties. The HttpHeaders addition has been moved from logging interceptors to separate contextHeadersDisabled-enabled interceptors in both configurations, which are sorted using a new OrderedComparator. Additionally, the 'depVersion' has been upgraded in the Gradle files, and a safer way of bean creation was introduced for 'HeaderHelper' in 'FilterConfig.kt'.