TIMEOFF-75 Remove redundant files and add comments to classes
The redundant files and classes within the project have been removed, including `report-group-plugin` directory and `AmqpRepeatableException.kt`. This lightens the codebase, allowing for easier navigation and maintenance. Comments explaining the usage and purpose of certain classes in the encryption and database packages were also added to improve code readability and context understanding for other developers.