Update dependency version and implement ordering in filters
The dependency version in the build files has been updated from 4.0.12 to 4.0.13. Additionally, the router filters now implement ordering. The ordering function has been added to the CaughtExceptionRouterFilter and LoggingConfigRouterFilter, while an Ordered import has been included in the RouterFilter class. These changes contribute to creating a logical and sequential procession of the filters.