Update dependency versions and introduce AlcosiHttpLoggingHelper
This update upgrades various dependencies in the build files from their previous versions to more recent ones for compatibility. It also introduces a new file, the AlcosiHttpLoggingHelper, which was created for logging HTTP Helper. Additionally, modifications were done in OKLoggingInterceptor and RestTemplateLogRequestResponseFilter to integrate AlcosiHttpLoggingHelper. These changes aim to improve the robustness and efficiency of the system.