Update dependency version and enhance logging configuration
The dependency version in 'build.gradle.kts' and 'with-dependency/build.gradle.kts' has been upgraded from 5.0.1 to 5.0.2. Additionally, the 'LoggingFilterConfig.kt' and 'LoggingFilterProperties.kt' classes were updated to enhance logging configuration. The former is now set to auto-configure prior to 'ServletLoggerConfiguration', and the latter has been marked as the primary configuration property.